Refactor tests to separate module
Moves the tests from `src/mem_fuse.rs` to a new file `src/tests/mirror_tests.rs`.
The tests are included in `src/main.rs` under a `#[cfg(test)]` flag.

To allow the tests to compile, several methods and the `nodes` field of `MemoryFuse` are made `pub(crate)`.
